# api.yaml
openapi: "3.0.1"
info:
  title: "OneHundredLettersApi"
  description: "API for managing correspondence, recipients and letters."
  version: "2025-03-11T16:08:24Z"
servers:
  - url: https://api.onehundredletters.com
    description: Production API
  - url: https://api-dev.onehundredletters.com
    description: Development API
security:
  - CognitoAuthorizer: []
paths:
...
  /recipient/{id}:
    get:
      summary: "Get Recipient By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the recipient to fetch."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Get Recipient By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/RecipientResponse"
                  message:
                    type: "string"
                    example: "Recipient fetched successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    put:
      summary: "Update Recipient By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the recipient to update."
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipientUpdateRequest"
        required: true
      responses:
        200:
          description: "Update Recipient By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    $ref: "#/components/schemas/RecipientResponse"
                  message:
                    type: "string"
                    example: "Recipient updated successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Recipient Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: "Delete Recipient By ID"
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "The ID of the recipient to delete."
          schema:
            type: "string"
            format: "uuid"
      responses:
        200:
          description: "Delete Recipient By ID Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "object"
                    properties:
                      recipientId:
                        $ref: "#/components/schemas/UUID"
                  message:
                    type: "string"
                    example: "Recipient deleted successfully!"
        400:
          description: "Bad Request Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        404:
          description: "Not Found Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
...
  schemas:
    Address:
      type: "object"
      properties:
        street:
          $ref: "#/components/schemas/Street"
        city:
          $ref: "#/components/schemas/City"
        state:
          $ref: "#/components/schemas/State"
        postalCode:
          $ref: "#/components/schemas/PostalCode"
        country:
          $ref: "#/components/schemas/Country"
      required:
        - street
        - city
        - state
        - postalCode
        - country
      example:
        street: "5623 Sunset Blvd"
        city: "Los Angeles"
        state: "CA"
        postalCode: "90047"
        country: "USA"
...
    City:
      type: "string"
      example: "Los Angeles"
...
    Country:
      type: "string"
      example: "USA"
    DateTime:
      type: "string"
      format: "date-time"
      example: "2025-03-07T14:30:00Z"
...
    FirstName:
      type: "string"
      example: "Alan"
...
    LastName:
      type: "string"
      example: "Aldrich"
...
    Occupation:
      type: "string"
      example: "Architect"
    Organization:
      type: "string"
      example: "Sera Architects"
    PostalCode:
      type: "string"
      example: "90047"
...
    RecipientDescription:
      type: "string"
      example: "George Church's contribution to the research community is invaluable."
    State:
      type: "string"
      example: "CA"
    Street:
      type: "string"
      example: "5623 Sunset Blvd"
...
    UUID:
      type: "string"
      format: "uuid"
      example: "52b249d0-7b41-4f58-bf49-d72d592f84ff"
...
    RecipientBase:
      type: "object"
      properties:
        address:
          $ref: "#/components/schemas/Address"
        firstName:
          $ref: "#/components/schemas/FirstName"
        lastName:
          $ref: "#/components/schemas/LastName"
        description:
          $ref: "#/components/schemas/RecipientDescription"
        occupation:
          $ref: "#/components/schemas/Occupation"
        organization:
          $ref: "#/components/schemas/Organization"
    RecipientResponse:
      allOf:
        - $ref: "#/components/schemas/RecipientBase"
        - type: "object"
          properties:
            recipientId:
              $ref: "#/components/schemas/UUID"
            createdAt:
              $ref: "#/components/schemas/DateTime"
            updatedAt:
              $ref: "#/components/schemas/DateTime"
...
    RecipientUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/RecipientBase"
        - type: "object"
          required:
            - "address"
            - "firstName"
            - "lastName"
            - "recipientId"
          properties:
            recipientId:
              $ref: "#/components/schemas/UUID"
...
    BadRequestError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 400
        message:
          type: "string"
          example: "Bad Request"
    InternalServerError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 500
        message:
          type: "string"
          example: "Internal Server Error"
    NotFoundError:
      type: "object"
      properties:
        statusCode:
          type: "integer"
          example: 404
        message:
          type: "string"
          example: "Not Found"
  securitySchemes:
    CognitoAuthorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      description: >-
        Bearer access token. Cognito User Pool issues tokens; API Gateway validates
        them via the sibling custom TOKEN Lambda authorizer (CFN AuthorizationType CUSTOM).
        Scheme name retained for compatibility; this is not a Cognito user-pools authorizer.
      x-amazon-apigateway-authtype: "custom"